@@ -1,6 +1,6 @@
-
-import { getConnection } from "@/lib/db";
+import { createDatabaseConnection } from "@/lib/db";
 import sql from 'mssql';
+import { getCached, setCached } from '@/lib/cache'; // ...existing code...
 
 function formatForFullTextSearch(text) {
   // Escape double quotes for SQL
@@ -17,12 +17,18 @@ export async function GET(req) {
   const company = searchParams.get("company")?.trim() || "";
   const sinceDate = searchParams.get("sinceDate")?.trim() || "";
 
+
+  const cachedRecentJobsCount = getCached('recent-jobs-count', { title, experienceLevel, location, company, sinceDate });
+  if (cachedRecentJobsCount) {
+    return new Response(JSON.stringify({ totalJobs: cachedRecentJobsCount }), { status: 200 });
+  }
+
   if (!sinceDate) {
     return new Response(JSON.stringify({ error: "sinceDate parameter is required" }), { status: 400 });
   }
 
   try {
-    const pool = await getConnection();
+    const pool = await createDatabaseConnection();
 
     // Build the count query without JOIN
     let query = `
@@ -37,22 +43,20 @@ export async function GET(req) {
     if (location) query += ` AND CONTAINS(jp.location, @location)`;
     if (company) query += ` AND jp.company_id = @company_id`;
 
-    const request = pool.request();
+    // Consolidate parameters
+    let params = {
+      sinceDate: new Date(sinceDate),
+    };
 
-    // Add parameters with proper formatting for full-text search
-    request.input('sinceDate', sql.DateTime, new Date(sinceDate));
-    if (title) request.input('title', sql.NVarChar, formatForFullTextSearch(title));
-    if (location) request.input('location', sql.NVarChar, formatForFullTextSearch(location));
-    if (experienceLevel) request.input('experienceLevel', sql.NVarChar, experienceLevel);
+    if (title) params.title = formatForFullTextSearch(title);
+    if (experienceLevel) params.experienceLevel = experienceLevel;
+    if (location) params.location = formatForFullTextSearch(location);
     if (company) {
       // If filtering by company, map company name to company_id
-      const companyQuery = `
+      const companyResult = await pool.executeQuery(`
         SELECT id FROM companies WITH (NOLOCK)
         WHERE LOWER(name) = LOWER(@companyName) AND deleted = 0
-      `;
-      const companyRequest = pool.request();
-      companyRequest.input('companyName', sql.NVarChar, company);
-      const companyResult = await companyRequest.query(companyQuery);
+      `, { companyName: company });
 
       if (companyResult.recordset.length === 0) {
         // No matching company found, return totalJobs as 0
@@ -60,11 +64,13 @@ export async function GET(req) {
       }
 
       const companyId = companyResult.recordset[0].id;
-      request.input('company_id', sql.Int, companyId);
+      params.company_id = companyId;
     }
 
-    const result = await request.query(query);
+    const result = await pool.executeQuery(query, params);
     const totalJobs = result.recordset[0]?.totalJobs || 0;
+
+    setCached('recent-jobs-count', { title, experienceLevel, location, company, sinceDate }, totalJobs);
 
     return new Response(JSON.stringify({ totalJobs }), { status: 200 });
   } catch (error) {
