@@ -77,83 +77,38 @@ import Link from "next/link";
 const footerLinks = [
   {
     href: "/job-postings",
-    icon: <BriefcaseBusiness size={16} />,
+    icon: <BriefcaseBusiness className="text-muted-foreground" size={14} />,
     text: "Job Postings",
   },
   {
-    href: "/job-postings?explevel=entry%20level",
-    icon: <User size={16} />,
+    href: "/job-postings?explevel=entry",
+    icon: <User className="text-muted-foreground" size={14} />,
     text: "Entry Level Jobs",
   },
   {
     href: "/job-postings?location=California",
-    icon: <Pin size={16} />,
+    icon: <Pin className="text-muted-foreground" size={14} />,
     text: "Jobs in California",
   },
   {
     href: "/job-postings?location=New%20York",
-    icon: <Pin size={16} />,
+    icon: <Pin className="text-muted-foreground" size={14} />,
     text: "Jobs in New York",
   },
   {
     href: "/job-postings?title=Software%20Engineer",
-    icon: <Code size={16} />,
+    icon: <Code className="text-muted-foreground" size={14} />,
     text: "Software Engineer Jobs",
-  },
-  {
-    href: "/job-postings?title=Project%20Manager",
-    icon: <Code size={16} />,
-    text: "Project Manager Jobs",
-  },
-  {
-    href: "/job-postings?explevel=mid%20level",
-    icon: <User size={16} />,
-    text: "Mid-Level Jobs",
-  },
-  {
-    href: "/job-postings?explevel=senior",
-    icon: <User size={16} />,
-    text: "Senior Level Jobs",
-  },
-  {
-    href: "/job-postings?location=Texas",
-    icon: <Pin size={16} />,
-    text: "Jobs in Texas",
-  },
-  {
-    href: "/job-postings?location=Florida",
-    icon: <Pin size={16} />,
-    text: "Jobs in Florida",
-  },
-  {
-    href: "/job-postings?title=Data%20Scientist",
-    icon: <Code size={16} />,
-    text: "Data Scientist Jobs",
-  },
-  {
-    href: "/job-postings?title=Product%20Manager",
-    icon: <Code size={16} />,
-    text: "Product Manager Jobs",
-  },
-  {
-    href: "/job-postings?company=86",
-    icon: <Code size={16} />,
-    text: "Microsoft Jobs",
-  },
-  {
-    href: "/job-postings?company=421",
-    icon: <Code size={16} />,
-    text: "Anthropic Jobs",
   },
 ];
 
 export default function Footer() {
   return (
-    <footer className="row-start-3 flex gap-6 mb-6 flex-wrap items-center justify-center">
-      <div className="flex px-4 flex-wrap gap-0 max-w-4xl justify-between gap-y-8">
+    <footer className="row-start-3 flex gap-3 mb-6 flex-wrap items-center justify-center">
+      <div className="flex px-4 flex-wrap gap-0 max-w-4xl justify-between gap-y-4">
         {footerLinks.map(({ href, icon, text }) => (
           <Link key={href} href={href}>
-            <p className="flex items-center gap-2 hover:underline hover:underline-offset-4">
+            <p className="flex items-center text-xs gap-2 hover:underline hover:underline-offset-4">
               {icon}
               {text}
             </p>
