@@ -84,9 +84,10 @@ export default function Button24({ jobId }) {
                     <div
                         onMouseEnter={() => setIsHovered(true)}
                         onMouseLeave={() => setIsHovered(false)}
+                        className="text-muted-foreground"
                     >
                         <Toggle
-                            className="group size-9 border shadow-sm hover:bg-green-500/10 hover:text-green-600 data-[state=on]:border-green-500/20 data-[state=on]:bg-green-500/10 data-[state=on]:text-green-600 dark:data-[state=on]:bg-green-500/10 dark:data-[state=on]:text-green-500"
+                            className="group size-9 border rounded-lg shadow-sm hover:bg-green-500/10 hover:text-green-600 data-[state=on]:border-green-500/20 data-[state=on]:bg-green-500/10 data-[state=on]:text-green-600 dark:data-[state=on]:bg-green-500/10 dark:data-[state=on]:text-green-500"
                             aria-label="Bookmark this"
                             pressed={bookmarked || isHovered}
                             onPressedChange={handleToggle}
@@ -95,7 +96,7 @@ export default function Button24({ jobId }) {
                                 size={16} 
                                 strokeWidth={2} 
                                 aria-hidden="true"
-                                className={isHovered ? "fill-current" : bookmarked ? "fill-current" : ""}
+                                className={isHovered ? "fill-current " : bookmarked ? "fill-current" : ""}
                             />
                         </Toggle>
                     </div>
