@@ -1,12 +1,13 @@
 import React from 'react';
 import Link from 'next/link';
+import Image from 'next/image';
 
 export function JobCard({ job }) {
     return (
         <div className="border rounded-lg p-4 mb-4 shadow-sm">
             <div className="flex items-center">
                 {job.logo && (
-                    <img src={job.logo} alt={`${job.company} logo`} className="w-12 h-12 mr-4" />
+                    <Image src={job.logo} alt={`${job.company} logo`} className="w-12 h-12 mr-4" />
                 )}
                 <div>
                     <h2 className="text-xl font-semibold">
