@@ -47,13 +47,16 @@ PaginationLink.displayName = "PaginationLink"
 
 const PaginationPrevious = ({
   className,
+  disabled,
   ...props
 }) => (
   <PaginationLink
     aria-label="Go to previous page"
     size="default"
-    className={cn("gap-1 pl-2.5", className)}
-    {...props}>
+    className={cn("gap-1 pl-2.5", className, disabled && "cursor-not-allowed opacity-50")}
+    onClick={disabled ? undefined : props.onClick}
+    {...props}
+  >
     <ChevronLeft className="h-4 w-4" />
     <span>Previous</span>
   </PaginationLink>
@@ -62,13 +65,16 @@ PaginationPrevious.displayName = "PaginationPrevious"
 
 const PaginationNext = ({
   className,
+  disabled,
   ...props
 }) => (
   <PaginationLink
     aria-label="Go to next page"
     size="default"
-    className={cn("gap-1 pr-2.5", className)}
-    {...props}>
+    className={cn("gap-1 pr-2.5", className, disabled && "cursor-not-allowed opacity-50")}
+    onClick={disabled ? undefined : props.onClick}
+    {...props}
+  >
     <span>Next</span>
     <ChevronRight className="h-4 w-4" />
   </PaginationLink>
